perf(card): split alarm chip/badge into their own module for fast cold-load render
The Verisure alarm "chip"/"badge" (custom elements `verisure-owa-alarm-chip`
/ `-badge`) was defined in the 90 KB `verisure-owa-alarm-card.js`. On a cold
companion-app load the whole bundle had to download before the element
registered with `customElements.define`, so the always-visible alarm chip
took 5-10s to appear — long enough that the alarm could trigger before the
user could tap to disarm. The entity state was fine throughout; this was
purely a frontend resource-load latency.

Split the monolith into three ES modules:
- verisure-owa-alarm-shared.js  - shared constants/helpers/translations
- verisure-owa-alarm-chip.js    - the chip + badge elements (lightweight)
- verisure-owa-alarm-card.js    - the heavy card + editor (slimmed 90->50 KB)

The chip module is registered as its own Lovelace resource, so the chip's
render path is now ~43 KB (chip + shared + utils) and loads independently of
the card+editor instead of being blocked by them. The moved class bodies are
byte-identical; only import/export wiring changed. card.js re-exports the
public helper API so existing imports keep working; the legacy
securitas-alarm-card.js shim imports both modules so a single legacy resource
still defines every element.

Also: when the popup is opened before the (separate) card module has loaded,
fall back to HA's native more-info dialog so the user can still arm/disarm
instead of the popup throwing.

Note: deliberately kept cache_headers=False on the static paths. A long
max-age would be safe for the hash-busted entry points but would serve the
bare-imported shared modules (no ?v=) stale for up to 31 days after an update;
a regression test documents this. Hard-caching is a possible follow-up once
the bare imports are cache-busted.
