+++ custom_components/securitas/const.py
@@ -37,6 +37,11 @@ def _card_url(filename: str) -> str:
 
 CARD_BASE_URL = "/verisure-owa-panel/verisure-owa-alarm-card.js"
 CARD_URL = _card_url("verisure-owa-alarm-card.js")
+# Lightweight chip/badge module — registered as its own resource so the
+# always-visible alarm chip renders without first downloading the heavy
+# alarm-card bundle.
+CHIP_CARD_BASE_URL = "/verisure-owa-panel/verisure-owa-alarm-chip.js"
+CHIP_CARD_URL = _card_url("verisure-owa-alarm-chip.js")
 CAMERA_CARD_BASE_URL = "/verisure-owa-panel/verisure-owa-camera-card.js"
 CAMERA_CARD_URL = _card_url("verisure-owa-camera-card.js")
 ACTIVITY_LOG_CARD_BASE_URL = "/verisure-owa-panel/verisure-owa-activity-log-card.js"
