+++ custom_components/securitas/www/verisure-owa-alarm-card.js
@@ -21,445 +21,42 @@
  *   name: My Alarm          # optional — overrides friendly_name
  */
 
-import { escHtml, formatTranslation } from "./verisure-owa-card-utils.js";
+import { escHtml } from "./verisure-owa-card-utils.js";
+import {
+  _t,
+  STATE_CFG,
+  STATE_COLOR_DEFAULTS,
+  COLOR_EDITOR_STATES,
+  STATE_LABEL_KEYS,
+  INACTIVE_STATES,
+  ARM_ACTIONS,
+  GESTURE_KEYS,
+  _filteredArmActions,
+  defaultArmState,
+  attachGesture,
+  TRANSLATIONS,
+  alarmEntitySuggestion,
+  _makeLegacyShim,
+} from "./verisure-owa-alarm-shared.js";
 
-// ── AlarmControlPanelEntityFeature bitmask values ────────────────────────────
-export const FEATURE = {
-  ARM_HOME: 1,
-  ARM_AWAY: 2,
-  ARM_NIGHT: 4,
-  ARM_CUSTOM_BYPASS: 16,
-  ARM_VACATION: 32,
-};
+// Re-export the public helper API so existing imports of these names from
+// this module keep working.
+export {
+  FEATURE,
+  TRANSLATIONS,
+  ARM_ACTIONS,
+  defaultArmState,
+  alarmEntitySuggestion,
+} from "./verisure-owa-alarm-shared.js";
 
-// ── Translations ─────────────────────────────────────────────────────────────
-export const TRANSLATIONS = {
-  en: {
-    disarmed: "Disarmed", armed_away: "Armed Away", armed_home: "Armed Home",
-    armed_night: "Armed Night", armed_vacation: "Armed Vacation",
-    armed_custom: "Armed Custom",
-    arming: "Arming\u2026", pending: "Pending", triggered: "TRIGGERED",
-    unavailable: "Unavailable", unknown: "Unknown",
-    arm_away: "Arm Away", arm_home: "Arm Home", arm_night: "Arm Night",
-    arm_vacation: "Arm Vacation", arm_custom: "Arm Custom", disarm: "Disarm",
-    force_arm: "Force Arm", cancel: "Cancel",
-    open_sensors: "Open sensor(s) \u2014 arm anyway?",
-    enter_pin: "Enter PIN to {action}", enter_code: "Enter code to {action}",
-    code: "Code", confirm: "Confirm",
-    refresh: "Refresh status",
-    waf_blocked: "Rate limited by Verisure servers. Please wait a few minutes.",
-    refresh_failed: "Refresh timed out — data may be stale.",
-    entity_not_found: "Entity not found: {entity}",
-    editor_entity: "Entity", editor_select: "\u2014 Select alarm panel \u2014",
-    editor_name: "Name (optional)", editor_name_placeholder: "Override friendly name",
-    editor_arm_modes: "Arm modes",
-    editor_arm_modes_hint: "Uncheck modes to hide their buttons. Modes not supported by the entity are not shown.",
-    editor_arm_modes_empty: "This entity reports no supported arm modes.",
-    editor_arm_state_no_modes: "Enable at least one arm mode above to set this action.",
-    card_name: "Verisure OWA Alarm Card",
-    card_description: "Alarm card for Verisure: dynamic arm modes, PIN support, force-arm for open sensors.",
-  },
-  es: {
-    disarmed: "Desarmado", armed_away: "Armado (fuera)", armed_home: "Armado (casa)",
-    armed_night: "Armado (noche)", armed_vacation: "Armado (vacaciones)",
-    armed_custom: "Armado (personalizado)",
-    arming: "Armando\u2026", pending: "Pendiente", triggered: "ACTIVADA",
-    unavailable: "No disponible", unknown: "Desconocido",
-    arm_away: "Armar fuera", arm_home: "Armar casa", arm_night: "Armar noche",
-    arm_vacation: "Armar vacaciones", arm_custom: "Armar personalizado", disarm: "Desarmar",
-    force_arm: "Forzar armado", cancel: "Cancelar",
-    open_sensors: "Sensor(es) abierto(s) \u2014 \u00bfarmar igualmente?",
-    enter_pin: "Introduzca PIN para {action}", enter_code: "Introduzca c\u00f3digo para {action}",
-    code: "C\u00f3digo", confirm: "Confirmar",
-    refresh: "Actualizar estado",
-    waf_blocked: "Bloqueado por los servidores de Verisure. Espere unos minutos.",
-    refresh_failed: "La actualización ha caducado — los datos pueden estar desactualizados.",
-    entity_not_found: "Entidad no encontrada: {entity}",
-    editor_entity: "Entidad", editor_select: "\u2014 Seleccionar panel de alarma \u2014",
-    editor_name: "Nombre (opcional)", editor_name_placeholder: "Nombre personalizado",
-    editor_arm_modes: "Modos de armado",
-    editor_arm_modes_hint: "Desmarque los modos para ocultar sus botones. Los modos no admitidos por la entidad no se muestran.",
-    editor_arm_modes_empty: "Esta entidad no admite ningún modo de armado.",
-    editor_arm_state_no_modes: "Habilite al menos un modo de armado arriba para configurar esta acción.",
-    card_name: "Tarjeta de Alarma Verisure",
-    card_description: "Tarjeta de alarma para Verisure: modos de armado, PIN y armado forzado.",
-  },
-  fr: {
-    disarmed: "D\u00e9sarm\u00e9", armed_away: "Arm\u00e9 (absent)", armed_home: "Arm\u00e9 (domicile)",
-    armed_night: "Arm\u00e9 (nuit)", armed_vacation: "Arm\u00e9 (vacances)",
-    armed_custom: "Arm\u00e9 (personnalis\u00e9)",
-    arming: "Armement\u2026", pending: "En attente", triggered: "D\u00c9CLENCH\u00c9E",
-    unavailable: "Indisponible", unknown: "Inconnu",
-    arm_away: "Armer absent", arm_home: "Armer domicile", arm_night: "Armer nuit",
-    arm_vacation: "Armer vacances", arm_custom: "Armer personnalis\u00e9", disarm: "D\u00e9sarmer",
-    force_arm: "Forcer l\u2019armement", cancel: "Annuler",
-    open_sensors: "Capteur(s) ouvert(s) \u2014 armer quand m\u00eame\u00a0?",
-    enter_pin: "Entrez le PIN pour {action}", enter_code: "Entrez le code pour {action}",
-    code: "Code", confirm: "Confirmer",
-    refresh: "Actualiser le statut",
-    waf_blocked: "Bloqu\u00e9 par les serveurs Verisure. Veuillez patienter quelques minutes.",
-    refresh_failed: "L\u2019actualisation a expir\u00e9 — les donn\u00e9es peuvent \u00eatre obsol\u00e8tes.",
-    entity_not_found: "Entit\u00e9 introuvable\u00a0: {entity}",
-    editor_entity: "Entit\u00e9", editor_select: "\u2014 S\u00e9lectionner le panneau d\u2019alarme \u2014",
-    editor_name: "Nom (facultatif)", editor_name_placeholder: "Remplacer le nom",
-    editor_arm_modes: "Modes d’armement",
-    editor_arm_modes_hint: "D\u00e9cochez les modes pour masquer leurs boutons. Les modes non pris en charge par l’entit\u00e9 ne sont pas affich\u00e9s.",
-    editor_arm_modes_empty: "Cette entit\u00e9 ne prend en charge aucun mode d\u2019armement.",
-    editor_arm_state_no_modes: "Activez au moins un mode d\u2019armement ci-dessus pour configurer cette action.",
-    card_name: "Carte d\u2019alarme Verisure",
-    card_description: "Carte d\u2019alarme Verisure\u00a0: modes d\u2019armement, PIN et armement forc\u00e9.",
-  },
-  it: {
-    disarmed: "Disarmato", armed_away: "Armato (fuori)", armed_home: "Armato (casa)",
-    armed_night: "Armato (notte)", armed_vacation: "Armato (vacanza)",
-    armed_custom: "Armato (personalizzato)",
-    arming: "Armamento\u2026", pending: "In attesa", triggered: "ATTIVATO",
-    unavailable: "Non disponibile", unknown: "Sconosciuto",
-    arm_away: "Arma fuori", arm_home: "Arma casa", arm_night: "Arma notte",
-    arm_vacation: "Arma vacanza", arm_custom: "Arma personalizzato", disarm: "Disarma",
-    force_arm: "Forza armamento", cancel: "Annulla",
-    open_sensors: "Sensore/i aperto/i \u2014 armare comunque?",
-    enter_pin: "Inserisci PIN per {action}", enter_code: "Inserisci codice per {action}",
-    code: "Codice", confirm: "Conferma",
-    refresh: "Aggiorna stato",
-    waf_blocked: "Bloccato dai server Verisure. Attendere qualche minuto.",
-    refresh_failed: "Aggiornamento scaduto — i dati potrebbero non essere aggiornati.",
-    entity_not_found: "Entit\u00e0 non trovata: {entity}",
-    editor_entity: "Entit\u00e0", editor_select: "\u2014 Seleziona pannello allarme \u2014",
-    editor_name: "Nome (facoltativo)", editor_name_placeholder: "Nome personalizzato",
-    editor_arm_modes: "Modalità di armamento",
-    editor_arm_modes_hint: "Deseleziona le modalità per nasconderne i pulsanti. Le modalità non supportate dall\u2019entità non vengono mostrate.",
-    editor_arm_modes_empty: "Questa entità non supporta alcuna modalità di armamento.",
-    editor_arm_state_no_modes: "Abilita almeno una modalità di armamento sopra per configurare questa azione.",
-    card_name: "Scheda Allarme Verisure",
-    card_description: "Scheda allarme Verisure: modalit\u00e0 di armamento, PIN e armamento forzato.",
-  },
-  pt: {
-    disarmed: "Desarmado", armed_away: "Armado (aus\u00eancia)", armed_home: "Armado (casa)",
-    armed_night: "Armado (noite)", armed_vacation: "Armado (f\u00e9rias)",
-    armed_custom: "Armado (personalizado)",
-    arming: "A armar\u2026", pending: "Pendente", triggered: "DISPARADO",
-    unavailable: "Indispon\u00edvel", unknown: "Desconhecido",
-    arm_away: "Armar aus\u00eancia", arm_home: "Armar casa", arm_night: "Armar noite",
-    arm_vacation: "Armar f\u00e9rias", arm_custom: "Armar personalizado", disarm: "Desarmar",
-    force_arm: "For\u00e7ar armamento", cancel: "Cancelar",
-    open_sensors: "Sensor(es) aberto(s) \u2014 armar na mesma?",
-    enter_pin: "Introduza PIN para {action}", enter_code: "Introduza c\u00f3digo para {action}",
-    code: "C\u00f3digo", confirm: "Confirmar",
-    refresh: "Atualizar estado",
-    waf_blocked: "Bloqueado pelos servidores da Verisure. Aguarde alguns minutos.",
-    refresh_failed: "A atualiza\u00e7\u00e3o expirou — os dados podem estar desatualizados.",
-    entity_not_found: "Entidade n\u00e3o encontrada: {entity}",
-    editor_entity: "Entidade", editor_select: "\u2014 Selecionar painel de alarme \u2014",
-    editor_name: "Nome (opcional)", editor_name_placeholder: "Nome personalizado",
-    editor_arm_modes: "Modos de armar",
-    editor_arm_modes_hint: "Desmarque os modos para ocultar os seus bot\u00f5es. Os modos n\u00e3o suportados pela entidade n\u00e3o s\u00e3o mostrados.",
-    editor_arm_modes_empty: "Esta entidade n\u00e3o suporta nenhum modo de armar.",
-    editor_arm_state_no_modes: "Ative pelo menos um modo de armar acima para configurar esta a\u00e7\u00e3o.",
-    card_name: "Cart\u00e3o de Alarme Verisure",
-    card_description: "Cart\u00e3o de alarme Verisure: modos de armar, PIN e armamento for\u00e7ado.",
-  },
-};
-
-// pt-BR falls back to pt
-TRANSLATIONS["pt-BR"] = TRANSLATIONS.pt;
-
-const _t = (lang, key, vars) => formatTranslation(lang, TRANSLATIONS, key, vars);
-
-// ── Per-state visual config ───────────────────────────────────────────────────
-const STATE_CFG = {
-  disarmed:           { icon: "mdi:shield-off-outline",  color: "var(--success-color,#4CAF50)" },
-  armed_away:         { icon: "mdi:shield-lock",         color: "var(--error-color,#F44336)" },
-  armed_home:         { icon: "mdi:shield-home",         color: "var(--warning-color,#FF9800)" },
-  armed_night:        { icon: "mdi:shield-moon",         color: "#9C27B0" },
-  armed_vacation:     { icon: "mdi:shield-airplane",     color: "#2196F3" },
-  armed_custom_bypass:{ icon: "mdi:shield-star",         color: "#00BCD4" },
-  arming:             { icon: "mdi:shield-sync-outline", color: "var(--warning-color,#FF9800)" },
-  pending:            { icon: "mdi:shield-alert-outline",color: "var(--warning-color,#FF9800)" },
-  triggered:          { icon: "mdi:shield-alert",        color: "var(--error-color,#F44336)" },
-  unavailable:        { icon: "mdi:shield-off-outline",  color: "var(--disabled-color,#9E9E9E)" },
-  unknown:            { icon: "mdi:shield-off-outline",  color: "var(--disabled-color,#9E9E9E)" },
-};
-
-// Hex fallbacks used by the editor color pickers (CSS vars can't be used in <input type="color">)
-const STATE_COLOR_DEFAULTS = {
-  disarmed:            "#4CAF50",
-  armed_away:          "#F44336",
-  armed_home:          "#FF9800",
-  armed_night:         "#9C27B0",
-  armed_vacation:      "#2196F3",
-  armed_custom_bypass: "#00BCD4",
-  triggered:           "#F44336",
-};
-
-// States shown in the color editor (excludes transient/unavailable states)
-const COLOR_EDITOR_STATES = [
-  { state: "disarmed",            label: "Disarmed" },
-  { state: "armed_away",          label: "Armed Away" },
-  { state: "armed_home",          label: "Armed Home" },
-  { state: "armed_night",         label: "Armed Night" },
-  { state: "armed_vacation",      label: "Armed Vacation" },
-  { state: "armed_custom_bypass", label: "Armed Custom / Bypass" },
-  { state: "triggered",           label: "Triggered" },
-];
-
-const STATE_LABEL_KEYS = {
-  disarmed: "disarmed", armed_away: "armed_away", armed_home: "armed_home",
-  armed_night: "armed_night", armed_vacation: "armed_vacation",
-  armed_custom_bypass: "armed_custom",
-  arming: "arming", pending: "pending", triggered: "triggered",
-  unavailable: "unavailable", unknown: "unknown",
-};
-
-// States where the alarm is considered armed (not disarmed/transitioning)
-const INACTIVE_STATES = new Set(["disarmed", "arming", "pending", "triggered", "unavailable", "unknown"]);
-
-// ── Arm action definitions ────────────────────────────────────────────────────
-export const ARM_ACTIONS = [
-  { key: "arm_away",          labelKey: "arm_away",    feature: FEATURE.ARM_AWAY,         service: "alarm_arm_away" },
-  { key: "arm_home",          labelKey: "arm_home",    feature: FEATURE.ARM_HOME,         service: "alarm_arm_home" },
-  { key: "arm_night",         labelKey: "arm_night",    feature: FEATURE.ARM_NIGHT,        service: "alarm_arm_night" },
-  { key: "arm_vacation",      labelKey: "arm_vacation", feature: FEATURE.ARM_VACATION,     service: "alarm_arm_vacation" },
-  { key: "arm_custom_bypass", labelKey: "arm_custom",   feature: FEATURE.ARM_CUSTOM_BYPASS,service: "alarm_arm_custom_bypass" },
-];
-
-const GESTURE_KEYS = ["tap_action", "hold_action", "double_tap_action"];
-
-// ── Gesture helpers ───────────────────────────────────────────────────────────
-
-/**
- * Returns `{ supported, filtered }` for an entity's arm capabilities.
- *  - `supported`: ARM_ACTIONS the entity advertises via `supported_features`.
- *  - `filtered`:  `supported` further intersected with `configStates` when
- *                 that array is provided; otherwise equal to `supported`.
- *
- * Used everywhere we need to honor both the entity's capabilities and the
- * user's optional `states` hide list.
- */
-function _filteredArmActions(features, configStates) {
-  const supported = ARM_ACTIONS.filter(a => features & a.feature);
-  const filtered = Array.isArray(configStates)
-    ? supported.filter(a => configStates.includes(a.key))
-    : supported;
-  return { supported, filtered };
-}
-
-/**
- * First arm state key the entity supports (and is in `configStates`, if
- * given), with fallbacks so callers always get a usable key.
- */
-export function defaultArmState(hass, entityId, configStates) {
-  const features = hass.states[entityId]?.attributes?.supported_features || 0;
-  const { supported, filtered } = _filteredArmActions(features, configStates);
-  const pool = filtered.length > 0 ? filtered : supported;
-  return pool.length > 0 ? pool[0].key : "arm_away";
-}
-
-// Entity registry platform this integration registers entities under. The
-// domain is `securitas` (a rename to `verisure_owa` was reversed before
-// release, so no install ever uses it).
-const _OUR_PLATFORMS = new Set(["securitas"]);
-
-// Card-picker suggestion hook: when the user selects one of our alarm panels,
-// offer both the full card and the compact chip as variants.
-export function alarmEntitySuggestion(hass, entityId) {
-  if (!entityId.startsWith("alarm_control_panel.")) return null;
-  if (!_OUR_PLATFORMS.has(hass?.entities?.[entityId]?.platform)) return null;
-  return [
-    { config: { type: "custom:verisure-owa-alarm-card", entity: entityId } },
-    { config: { type: "custom:verisure-owa-alarm-chip", entity: entityId } },
-  ];
-}
-
-/**
- * Attaches pointer-based gesture listeners to `el`.
- *
- * Gesture logic:
- *  - Long-press  : pointerdown → 500 ms timer. Cancel on >10 px move or
- *                  pointerup/cancel before timer fires. When timer fires,
- *                  executes hold_action and suppresses the next click.
- *  - Double-tap  : first pointerup starts a 300 ms window. Second pointerup
- *                  within the window executes double_tap_action. Window
- *                  expiry executes tap_action.
- *  - Single tap  : click event executes tap_action (unless suppressed by
- *                  long-press).
- *
- * @param {HTMLElement}   el          - Element to attach listeners to
- * @param {object}        config      - Card/badge config (tap_action etc.)
- * @param {object}        hass        - Home Assistant hass object
- * @param {string}        entityId    - Alarm entity id
- * @param {HTMLElement}   srcEl       - Element to dispatch events from
- * @param {object}        callbacks   - { startPinEntry(action), onMoreInfo() }
- * @param {string[]}      [cardStates] - Card's `_config.states` (optional);
- *                                       passed to executeAction so its
- *                                       arm_or_disarm fallback honors the
- *                                       user's filtered subset.
- * @returns {Function}                - Cleanup function (removes listeners)
- */
-function attachGesture(el, config, hass, entityId, srcEl, callbacks = {}, cardStates) {
-  let holdTimer = null;
-  let holdFired = false;
-  let downX = 0, downY = 0;
-  let tapWindow = null;
-
-
-  const HOLD_MS    = 500;
-  const DOUBLE_MS  = 300;
-  const MOVE_PX    = 10;
-
-  const tapAction       = config.tap_action       || { action: "more-info" };
-  const holdAction      = config.hold_action      || { action: "none" };
-  const doubleTapAction = config.double_tap_action || { action: "none" };
-
-  function cancelHold() {
-    if (holdTimer) { clearTimeout(holdTimer); holdTimer = null; }
-  }
-
-  function onPointerDown(e) {
-    holdFired = false;
-    downX = e.clientX; downY = e.clientY;
-    holdTimer = setTimeout(() => {
-      holdTimer = null;
-      holdFired = true;
-      executeAction(holdAction, hass, entityId, srcEl, callbacks, cardStates);
-    }, HOLD_MS);
-  }
-
-  function onPointerMove(e) {
-    if (holdTimer) {
-      const dx = e.clientX - downX, dy = e.clientY - downY;
-      if (Math.sqrt(dx * dx + dy * dy) > MOVE_PX) cancelHold();
-    }
-  }
-
-  function onPointerUp() {
-    cancelHold();
-    if (holdFired) return;
-
-    if (tapWindow) {
-      clearTimeout(tapWindow);
-      tapWindow = null;
-      executeAction(doubleTapAction, hass, entityId, srcEl, callbacks, cardStates);
-    } else {
-      tapWindow = setTimeout(() => {
-        tapWindow = null;
-        executeAction(tapAction, hass, entityId, srcEl, callbacks, cardStates);
-      }, DOUBLE_MS);
-    }
-  }
-
-  function onPointerCancel() { cancelHold(); }
-
-  function onClick(e) {
-    // Our gesture handler owns the click on this element. Always swallow
-    // the native click so it cannot bubble to a parent's tap_action handler
-    // (e.g. an HA tile-card wrapper or dashboard view default) — otherwise
-    // a single tap would fire both our action AND the parent's, opening
-    // duplicate dialogs.
-    if (holdFired) holdFired = false;
-    e.stopImmediatePropagation();
-  }
-
-  el.addEventListener("pointerdown",   onPointerDown);
-  el.addEventListener("pointermove",   onPointerMove);
-  el.addEventListener("pointerup",     onPointerUp);
-  el.addEventListener("pointercancel", onPointerCancel);
-  el.addEventListener("click",         onClick, true);
-
-  return function cleanup() {
-    el.removeEventListener("pointerdown",   onPointerDown);
-    el.removeEventListener("pointermove",   onPointerMove);
-    el.removeEventListener("pointerup",     onPointerUp);
-    el.removeEventListener("pointercancel", onPointerCancel);
-    el.removeEventListener("click",         onClick, true);
-    cancelHold();
-    if (tapWindow) { clearTimeout(tapWindow); tapWindow = null; }
-  };
-}
-
-/**
- * Executes a HA-style action config object.
- *
- * @param {object}      action      - { action, navigation_path, perform_action, data, arm_state }
- * @param {object}      hass        - Home Assistant hass object
- * @param {string}      entityId    - Alarm entity id
- * @param {HTMLElement} srcEl       - Element to dispatch events from (for more-info)
- * @param {object}      callbacks   - { startPinEntry(serviceAction), onMoreInfo() }
- * @param {string[]}    [cardStates] - Card's `_config.states` (optional); when
- *                                     `arm_or_disarm` has no explicit
- *                                     `arm_state`, the fallback default is
- *                                     drawn from this filtered subset.
- */
-function executeAction(action, hass, entityId, srcEl, callbacks = {}, cardStates) {
-  if (!action || action.action === "none") return;
-
-  switch (action.action) {
-
-    case "more-info":
-      if (callbacks.onMoreInfo) {
-        callbacks.onMoreInfo();
-      } else if (srcEl) {
-        srcEl.dispatchEvent(new CustomEvent("hass-more-info", {
-          detail: { entityId },
-          bubbles: true,
-          composed: true,
-        }));
-      }
-      break;
-
-    case "navigate": {
-      const path = action.navigation_path;
-      if (path) {
-        history.pushState({}, "", path);
-        window.dispatchEvent(new Event("location-changed"));
-      }
-      break;
-    }
-
-    case "perform-action": {
-      const call = action.perform_action || "";
-      const dot  = call.indexOf(".");
-      if (dot > 0) {
-        hass.callService(call.slice(0, dot), call.slice(dot + 1), action.data || {});
-      }
-      break;
-    }
-
-    case "arm_or_disarm": {
-      const stateObj = hass.states[entityId];
-      if (!stateObj) return;
-      const state          = stateObj.state;
-      const attrs          = stateObj.attributes;
-      const isArmed        = !INACTIVE_STATES.has(state);
-      const hasCode        = !!attrs.code_format;
-      const codeArmReq     = attrs.code_arm_required === true;
-
-      if (isArmed || state === "arming" || state === "pending" || state === "triggered") {
-        // Disarm
-        const svcAction = { service: "alarm_disarm", labelKey: "disarm" };
-        if (hasCode && callbacks.startPinEntry) {
-          callbacks.startPinEntry(svcAction);
-        } else {
-          hass.callService("alarm_control_panel", "alarm_disarm", { entity_id: entityId });
-        }
-      } else if (state === "disarmed") {
-        // Arm
-        const armKey = action.arm_state || defaultArmState(hass, entityId, cardStates);
-        const armDef = ARM_ACTIONS.find(a => a.key === armKey);
-        if (!armDef) return;
-        const svcAction = { service: armDef.service, labelKey: armDef.labelKey };
-        if (hasCode && codeArmReq && callbacks.startPinEntry) {
-          callbacks.startPinEntry(svcAction);
-        } else {
-          hass.callService("alarm_control_panel", armDef.service, { entity_id: entityId });
-        }
-      }
-      break;
-    }
-  }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
+// The lightweight chip/badge are defined in verisure-owa-alarm-chip.js, which
+// the integration registers as a SEPARATE Lovelace resource so the
+// always-visible alarm chip renders without waiting for this heavier card
+// bundle. It is deliberately NOT imported here: the relative import resolves to
+// a different URL than the hash-busted resource, so importing it would just
+// fetch chip.js a second time for no benefit. The badge/chip create the full
+// card lazily (document.createElement) when their popup opens — by then this
+// resource has loaded.
 
 class VerisureOwaAlarmCard extends HTMLElement {
   constructor() {
@@ -1674,460 +1271,14 @@ class VerisureOwaAlarmCardEditor extends HTMLElement {
   }
 }
 
-// ── Badge card ────────────────────────────────────────────────────────────────
-
-class VerisureOwaAlarmBadge extends HTMLElement {
-  constructor() {
-    super();
-    this.attachShadow({ mode: "open" });
-    this._dialogOpen = false;
-    this._pinOverlay = null;   // floating PIN overlay element (or null)
-    this._pinState   = null;   // { service, labelKey } when PIN entry active
-    this._pin        = "";
-    this._gestureCleanup = null; // cleanup fn returned by attachGesture
-  }
-
-  disconnectedCallback() {
-    if (this._gestureCleanup) { this._gestureCleanup(); this._gestureCleanup = null; }
-    if (this._pinOverlay) { this._pinOverlay.remove(); this._pinOverlay = null; }
-    // Reset for re-render on reconnection — see VerisureOwaAlarmCard.disconnectedCallback.
-    this._lastKey = null;
-    this._pinState = null;
-    this._pin = "";
-  }
-
-  setConfig(config) {
-    if (!config.entity) throw new Error("Please define an entity");
-    this._config = config;
-    this._lastKey = null;
-  }
-
-  set hass(hass) {
-    this._hass = hass;
-    const stateObj = hass.states[this._config.entity];
-    const newKey = stateObj
-      ? `${stateObj.state}|${stateObj.attributes.force_arm_available}`
-      : "missing";
-    if (newKey !== this._lastKey) {
-      this._lastKey = newKey;
-      this._renderBadge();
-    }
-    // Forward hass to the dialog card if open
-    if (this._dialogCard) this._dialogCard.hass = hass;
-  }
-
-  _renderBadge() {
-    if (!this._hass || !this._config) return;
-
-    const stateObj = this._hass.states[this._config.entity];
-    if (!stateObj) {
-      this.shadowRoot.innerHTML = `<ha-icon icon="mdi:shield-alert" style="color:var(--error-color)"></ha-icon>`;
-      return;
-    }
-
-    const state = stateObj.state;
-    const icons = stateObj.attributes.force_arm_available
-      ? { icon: "mdi:alert", color: "var(--warning-color, #FF9800)" }
-      : STATE_CFG[state] || { icon: "mdi:shield", color: "var(--disabled-color,#9E9E9E)" };
-
-    this.shadowRoot.innerHTML = `
-      <style>
-        :host { display: inline-block; }
-        .badge {
-          cursor: pointer;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          transition: transform 0.1s;
-        }
-        .badge:active { transform: scale(0.9); }
-        .badge ha-icon {
-          --mdc-icon-size: 24px;
-          color: ${icons.color};
-        }
-      </style>
-      <div class="badge" id="badge">
-        <ha-icon icon="${icons.icon}"></ha-icon>
-      </div>`;
-
-    // Clean up previous gesture listeners (badge re-renders on state change)
-    if (this._gestureCleanup) { this._gestureCleanup(); this._gestureCleanup = null; }
-
-    const badgeEl = this.shadowRoot.getElementById("badge");
-    const gestureConfig = {
-      tap_action:        this._config.tap_action        || { action: "more-info" },
-      hold_action:       this._config.hold_action       || { action: "arm_or_disarm", arm_state: defaultArmState(this._hass, this._config.entity, this._config.states) },
-      double_tap_action: this._config.double_tap_action || { action: "none" },
-    };
-
-    this._gestureCleanup = attachGesture(
-      badgeEl,
-      gestureConfig,
-      this._hass,
-      this._config.entity,
-      this,
-      {
-        onMoreInfo:    () => this._openDialog(),
-        startPinEntry: (svcAction) => this._startBadgePinEntry(svcAction),
-      },
-      this._config.states,
-    );
-  }
-
-  _startBadgePinEntry(svcAction) {
-    if (this._pinOverlay) return; // already showing
-
-    const hass   = this._hass;
-    const entity = this._config.entity;
-    const lang   = hass.language || "en";
-    const stateObj = hass.states[entity];
-    const codeFormat = stateObj?.attributes?.code_format || "number";
-
-    this._pinState = svcAction;
-    this._pin      = "";
-
-    const overlay = document.createElement("div");
-    Object.assign(overlay.style, {
-      position: "fixed", top: "0", left: "0", right: "0", bottom: "0",
-      background: "rgba(0,0,0,0.5)", zIndex: "8",
-      display: "flex", alignItems: "center", justifyContent: "center",
-      padding: "16px",
-    });
-
-    const box = document.createElement("div");
-    Object.assign(box.style, {
-      width: "100%", maxWidth: "340px",
-      borderRadius: "16px",
-      background: "var(--card-background-color, var(--ha-card-background, #fff))",
-      boxShadow: "0 8px 32px rgba(0,0,0,0.25)",
-      padding: "20px",
-      fontFamily: "inherit",
-    });
-
-    const actionLabel = svcAction.labelKey ? _t(lang, svcAction.labelKey) : (svcAction.label || "");
-    const promptKey   = codeFormat === "number" ? "enter_pin" : "enter_code";
-
-    box.innerHTML = `
-      <div style="font-size:0.9em;font-weight:600;color:var(--primary-text-color);margin-bottom:12px">
-        ${_t(lang, promptKey, { action: actionLabel })}
-      </div>
-      ${codeFormat === "number" ? `
-        <input id="badge-pin-input" type="password" inputmode="numeric" autocomplete="off"
-               style="width:100%;box-sizing:border-box;padding:8px 12px;border:1px solid var(--divider-color);
-                      border-radius:8px;font-size:1.1em;margin-bottom:12px;background:var(--secondary-background-color);
-                      color:var(--primary-text-color)" placeholder="••••" />
-        <div style="display:grid;grid-template-columns:repeat(3,1fr);gap:6px;margin-bottom:12px">
-          ${[1,2,3,4,5,6,7,8,9].map(n =>
-            `<button data-badge-key="${n}" style="padding:10px;border:none;border-radius:8px;font-size:1em;font-weight:600;cursor:pointer;background:var(--secondary-background-color);color:var(--primary-text-color)">${n}</button>`
-          ).join("")}
-          <button data-badge-key="cancel" style="padding:10px;border:none;border-radius:8px;font-size:1em;cursor:pointer;background:var(--secondary-background-color);color:var(--error-color)">✕</button>
-          <button data-badge-key="0" style="padding:10px;border:none;border-radius:8px;font-size:1em;font-weight:600;cursor:pointer;background:var(--secondary-background-color);color:var(--primary-text-color)">0</button>
-          <button data-badge-key="del" style="padding:10px;border:none;border-radius:8px;font-size:1em;cursor:pointer;background:var(--secondary-background-color);color:var(--primary-text-color)">⌫</button>
-        </div>
-      ` : `
-        <input id="badge-pin-input" type="password" autocomplete="off"
-               style="width:100%;box-sizing:border-box;padding:8px 12px;border:1px solid var(--divider-color);
-                      border-radius:8px;font-size:1em;margin-bottom:12px;background:var(--secondary-background-color);
-                      color:var(--primary-text-color)" placeholder="${_t(lang, "code")}" />
-      `}
-      <div style="display:flex;gap:8px">
-        <button id="badge-pin-cancel" style="flex:1;padding:10px;border:none;border-radius:8px;font-size:0.9em;font-weight:600;cursor:pointer;background:var(--secondary-background-color);color:var(--primary-text-color)">${_t(lang, "cancel")}</button>
-        <button id="badge-pin-confirm" style="flex:1;padding:10px;border:none;border-radius:8px;font-size:0.9em;font-weight:600;cursor:pointer;background:var(--primary-color);color:var(--text-primary-color,#fff)">${_t(lang, "confirm")}</button>
-      </div>`;
-
-    overlay.appendChild(box);
-    document.body.appendChild(overlay);
-    this._pinOverlay = overlay;
-
-    const close = () => {
-      overlay.remove();
-      this._pinOverlay = null;
-      this._pinState   = null;
-      this._pin        = "";
-    };
-
-    // Keypad
-    const pinInput = box.querySelector("#badge-pin-input");
-    const syncInput = () => { if (pinInput) pinInput.value = this._pin; };
-
-    box.querySelectorAll("[data-badge-key]").forEach(btn => {
-      btn.addEventListener("click", () => {
-        const k = btn.dataset.badgeKey;
-        if (k === "cancel") { close(); return; }
-        if (k === "del")    { this._pin = this._pin.slice(0, -1); syncInput(); return; }
-        this._pin += k; syncInput();
-      });
-    });
-
-    if (pinInput) {
-      requestAnimationFrame(() => pinInput.focus());
-      pinInput.addEventListener("input", e => {
-        this._pin = codeFormat === "number"
-          ? e.target.value.replace(/\D/g, "")
-          : e.target.value;
-        if (codeFormat === "number") e.target.value = this._pin;
-      });
-      pinInput.addEventListener("keydown", e => {
-        if (e.key === "Enter")  this._submitBadgePin(close);
-        if (e.key === "Escape") close();
-      });
-    }
-
-    box.querySelector("#badge-pin-cancel").addEventListener("click", close);
-    box.querySelector("#badge-pin-confirm").addEventListener("click", () => this._submitBadgePin(close));
-
-    // Tap outside to close
-    overlay.addEventListener("click", e => { if (e.target === overlay) close(); });
-  }
-
-  _submitBadgePin(closeFn) {
-    if (!this._pinState || !this._pin) return;
-    this._hass.callService("alarm_control_panel", this._pinState.service, {
-      entity_id: this._config.entity,
-      code: this._pin,
-    });
-    closeFn();
-  }
-
-  _openDialog() {
-    if (this._dialogOpen) return;
-    this._dialogOpen = true;
-
-    const overlay = document.createElement("div");
-    Object.assign(overlay.style, {
-      position: "fixed", top: "0", left: "0", right: "0", bottom: "0",
-      background: "rgba(0,0,0,0.5)", zIndex: "7",
-      display: "flex", alignItems: "center", justifyContent: "center",
-      padding: "16px",
-    });
-
-    const content = document.createElement("div");
-    Object.assign(content.style, {
-      width: "100%", maxWidth: "400px", maxHeight: "90vh", overflowY: "auto",
-      borderRadius: "16px", background: "var(--card-background-color, var(--ha-card-background, #fff))",
-      boxShadow: "0 8px 32px rgba(0,0,0,0.25)", position: "relative",
-    });
-
-    const closeBtn = document.createElement("button");
-    closeBtn.textContent = "\u2715";
-    Object.assign(closeBtn.style, {
-      position: "absolute", top: "8px", right: "8px", width: "32px", height: "32px",
-      border: "none", borderRadius: "50%", background: "var(--secondary-background-color)",
-      color: "var(--primary-text-color)", fontSize: "1.1em", cursor: "pointer",
-      display: "flex", alignItems: "center", justifyContent: "center", zIndex: "1",
-    });
-
-    content.appendChild(closeBtn);
-    overlay.appendChild(content);
-    document.body.appendChild(overlay);
-
-    // Create the full alarm card inside the dialog.
-    //
-    // Strip the badge/chip's gesture config before passing it down. The same
-    // tap_action key means different things in different contexts: on a badge
-    // or chip, `more-info` is wired to open THIS popup; on the alarm-card,
-    // `more-info` dispatches `hass-more-info` and opens HA's standard dialog.
-    // Forwarding the badge's gestures verbatim would make a tap on the icon
-    // inside the popup open HA's dialog on top of our popup.
-    const innerConfig = { ...this._config };
-    for (const k of GESTURE_KEYS) delete innerConfig[k];
-    this._dialogCard = document.createElement("securitas-alarm-card");
-    this._dialogCard.setConfig(innerConfig);
-    this._dialogCard.hass = this._hass;
-    content.appendChild(this._dialogCard);
-
-    // Close handlers
-    const close = () => {
-      this._dialogOpen = false;
-      this._dialogCard = null;
-      overlay.remove();
-      if (this._unsubConnection) {
-        this._unsubConnection();
-        this._unsubConnection = null;
-      }
-    };
-    closeBtn.addEventListener("click", close);
-    overlay.addEventListener("click", (e) => {
-      if (e.target === overlay) close();
-    });
-
-    // Close overlay when HA connection drops (e.g. restart)
-    if (this._hass?.connection) {
-      this._unsubConnection = this._hass.connection.addEventListener(
-        "disconnected", close
-      );
-    }
-  }
-
-  getCardSize() { return 1; }
-
-  static getConfigElement() {
-    return document.createElement("verisure-owa-alarm-card-editor");
-  }
-
-  static getStubConfig(hass) {
-    const entities = Object.keys(hass.states).filter(e => e.startsWith("alarm_control_panel."));
-    return { entity: entities[0] || "" };
-  }
-}
-
-// ── Mushroom-compatible chip ─────────────────────────────────────────────────
-
-class VerisureOwaAlarmChip extends HTMLElement {
-  constructor() {
-    super();
-    this.attachShadow({ mode: "open" });
-    this._dialogOpen = false;
-    this._pinOverlay = null;
-    this._pinState   = null;
-    this._pin        = "";
-    this._gestureCleanup = null;
-  }
-
-  disconnectedCallback() {
-    if (this._gestureCleanup) { this._gestureCleanup(); this._gestureCleanup = null; }
-    if (this._pinOverlay) { this._pinOverlay.remove(); this._pinOverlay = null; }
-    // Reset for re-render on reconnection — see VerisureOwaAlarmCard.disconnectedCallback.
-    this._lastKey = null;
-    this._pinState = null;
-    this._pin = "";
-  }
-
-  setConfig(config) {
-    if (!config.entity) throw new Error("Please define an entity");
-    this._config = config;
-    this._lastKey = null;  // force re-render on config change
-    if (this._hass) this._tryRender();
-  }
-
-  set config(config) { this.setConfig(config); }
-
-  set hass(hass) {
-    this._hass = hass;
-    this._tryRender();
-    if (this._dialogCard) this._dialogCard.hass = hass;
-  }
-
-  _tryRender() {
-    if (!this._hass || !this._config) return;
-    const stateObj = this._hass.states[this._config.entity];
-    const newKey = stateObj
-      ? `${stateObj.state}|${stateObj.attributes.force_arm_available}`
-      : "missing";
-    if (newKey !== this._lastKey) {
-      this._lastKey = newKey;
-      this._render();
-    }
-  }
-
-  _render() {
-    if (!this._hass || !this._config) return;
-
-    const stateObj = this._hass.states[this._config.entity];
-    if (!stateObj) {
-      this.shadowRoot.innerHTML = `<ha-icon icon="mdi:shield-alert" style="color:var(--error-color)"></ha-icon>`;
-      return;
-    }
-
-    const state = stateObj.state;
-    const cfg = stateObj.attributes.force_arm_available
-      ? { icon: "mdi:alert", color: "var(--warning-color, #FF9800)" }
-      : STATE_CFG[state] || { icon: "mdi:shield", color: "var(--disabled-color,#9E9E9E)" };
-
-    this.shadowRoot.innerHTML = `
-      <style>
-        :host {
-          display: flex;
-          --chip-height: 36px;
-          --chip-padding: 0 10px;
-          --chip-border-radius: 19px;
-          --chip-icon-size: 18px;
-        }
-        .chip {
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          height: var(--chip-height);
-          padding: var(--chip-padding);
-          border-radius: var(--chip-border-radius);
-          background: var(--ha-card-background, var(--card-background-color, #fff));
-          box-shadow: var(--chip-box-shadow, 0 2px 4px rgba(0,0,0,0.06));
-          cursor: pointer;
-          transition: transform 0.1s;
-          user-select: none;
-          -webkit-user-select: none;
-        }
-        .chip:active { transform: scale(0.95); }
-        .chip ha-icon {
-          --mdc-icon-size: var(--chip-icon-size);
-          color: ${cfg.color};
-        }
-      </style>
-      <div class="chip" id="chip">
-        <ha-icon icon="${cfg.icon}"></ha-icon>
-      </div>`;
-
-    if (this._gestureCleanup) { this._gestureCleanup(); this._gestureCleanup = null; }
-
-    const chipEl = this.shadowRoot.getElementById("chip");
-    const gestureConfig = {
-      tap_action:        this._config.tap_action        || { action: "more-info" },
-      hold_action:       this._config.hold_action       || { action: "none" },
-      double_tap_action: this._config.double_tap_action || { action: "none" },
-    };
-
-    this._gestureCleanup = attachGesture(
-      chipEl,
-      gestureConfig,
-      this._hass,
-      this._config.entity,
-      this,
-      {
-        onMoreInfo:    () => this._openDialog(),
-        startPinEntry: (svcAction) => VerisureOwaAlarmBadge.prototype._startBadgePinEntry.call(this, svcAction),
-      },
-      this._config.states,
-    );
-  }
-
-  _openDialog() {
-    VerisureOwaAlarmBadge.prototype._openDialog.call(this);
-  }
-
-  _submitBadgePin(closeFn) {
-    VerisureOwaAlarmBadge.prototype._submitBadgePin.call(this, closeFn);
-  }
-
-  getCardSize() { return 1; }
-}
-
 /* v8 ignore start -- defensive duplicate-registration guards;
    the "already defined" branches can't be hit in single-process tests. */
-
-// Canonical (post-v5) tags. After Phase H this is the name HACS will show in the picker.
 if (!customElements.get("verisure-owa-alarm-card")) {
   customElements.define("verisure-owa-alarm-card", VerisureOwaAlarmCard);
 }
 if (!customElements.get("verisure-owa-alarm-card-editor")) {
   customElements.define("verisure-owa-alarm-card-editor", VerisureOwaAlarmCardEditor);
 }
-if (!customElements.get("verisure-owa-alarm-badge")) {
-  customElements.define("verisure-owa-alarm-badge", VerisureOwaAlarmBadge);
-}
-if (!customElements.get("verisure-owa-alarm-chip")) {
-  customElements.define("verisure-owa-alarm-chip", VerisureOwaAlarmChip);
-}
-if (!customElements.get("mushroom-verisure-owa-alarm-chip")) {
-  customElements.define("mushroom-verisure-owa-alarm-chip", class extends VerisureOwaAlarmChip {});
-}
-
-// `securitas-*` tag-name aliases for the canonical `verisure-owa-*`
-// cards. Kept registered indefinitely — both names are equal-weight,
-// no deprecation. The card picker only offers the verisure-owa form.
-function _makeLegacyShim(canonicalClass, _oldTag, _newTag) {
-  return class extends canonicalClass {};
-}
-
 if (!customElements.get("securitas-alarm-card")) {
   customElements.define("securitas-alarm-card",
     _makeLegacyShim(VerisureOwaAlarmCard, "securitas-alarm-card", "verisure-owa-alarm-card"));
@@ -2135,18 +1286,6 @@ if (!customElements.get("securitas-alarm-card")) {
 if (!customElements.get("securitas-alarm-card-editor")) {
   customElements.define("securitas-alarm-card-editor",
     _makeLegacyShim(VerisureOwaAlarmCardEditor, "securitas-alarm-card-editor", "verisure-owa-alarm-card-editor"));
-}
-if (!customElements.get("securitas-alarm-badge")) {
-  customElements.define("securitas-alarm-badge",
-    _makeLegacyShim(VerisureOwaAlarmBadge, "securitas-alarm-badge", "verisure-owa-alarm-badge"));
-}
-if (!customElements.get("securitas-alarm-chip")) {
-  customElements.define("securitas-alarm-chip",
-    _makeLegacyShim(VerisureOwaAlarmChip, "securitas-alarm-chip", "verisure-owa-alarm-chip"));
-}
-if (!customElements.get("mushroom-securitas-alarm-chip")) {
-  customElements.define("mushroom-securitas-alarm-chip",
-    _makeLegacyShim(VerisureOwaAlarmChip, "mushroom-securitas-alarm-chip", "mushroom-verisure-owa-alarm-chip"));
 }
 
 window.customCards = window.customCards || [];
@@ -2157,23 +1296,6 @@ if (!window.customCards.find(c => c.type === "verisure-owa-alarm-card")) {
     description: TRANSLATIONS.en.card_description,
     preview:     false,
     getEntitySuggestion: alarmEntitySuggestion,
-  });
-}
-if (!window.customCards.find(c => c.type === "verisure-owa-alarm-chip")) {
-  window.customCards.push({
-    type:        "verisure-owa-alarm-chip",
-    name:        "Verisure OWA Alarm Chip",
-    description: "Mushroom-compatible alarm chip — shows alarm state with force-arm support.",
-    preview:     false,
-  });
-}
-window.customBadges = window.customBadges || [];
-if (!window.customBadges.find(b => b.type === "verisure-owa-alarm-badge")) {
-  window.customBadges.push({
-    type:        "verisure-owa-alarm-badge",
-    name:        "Verisure OWA Alarm Badge",
-    description: "Compact alarm badge — click to open full alarm card with force-arm support.",
-    preview:     false,
   });
 }
 /* v8 ignore stop */
