+++ tests-js/integration/legacy-shims.test.js
@@ -15,4 +15,13 @@ describe("legacy shim re-exports", () => {
     expect(customElements.get("securitas-alarm-card")).toBeDefined();
     expect(customElements.get("securitas-camera-card")).toBeDefined();
   });
+
+  it("loading securitas-alarm-card.js also registers the chip/badge (split modules)", () => {
+    // The chip/badge now live in a separate module; the shim imports it too so
+    // a single legacy resource still defines every element, as before the split.
+    expect(customElements.get("securitas-alarm-chip")).toBeDefined();
+    expect(customElements.get("securitas-alarm-badge")).toBeDefined();
+    expect(customElements.get("verisure-owa-alarm-chip")).toBeDefined();
+    expect(customElements.get("verisure-owa-alarm-badge")).toBeDefined();
+  });
 });
